+++ packages/db/src/search/field-mapping.ts
@@ -123,6 +123,18 @@ const collectionMapping : { [key: string]: Function } = {
   updatedAt: (value?: string) => ({
     updated_at: value && new Date(Number(value)),
   }),
+  reportToEms: (value?: string) => ({
+    report_to_ems: (value === 'true'),
+  }),
+  process: (value?: string) => ({
+    process: value,
+  }),
+  sampleFileName: (value?: string) => ({
+    sample_file_name: value,
+  }),
+  urlPath: (value?: string) => ({
+    url_path: value,
+  }),
 };
 
 const executionMapping : { [key: string]: Function } = {